% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/template.R
\name{template}
\alias{template}
\title{Template}
\usage{
template(num, type)
}
\arguments{
\item{num}{A number to learn about.}

\item{type}{The category in which to learn.}
}
\value{
The a fact about \code{num} in \code{type} category.
}
\description{
Template
}
\examples{
template_get(42, "math")

}
